scraper/requirements.txt
asyncpraw
supabase
openai
python-dotenv
